chore: move tests from module -> wasmer_sys to avoid public glob rexport shadowing
